\name{mgVarPart}
\alias{mgVarPart}
\title{
Variation partitioning using MEMGENE variables
}
\description{
This function wraps the vegan::varpart function to simplify variation partitioning using the MEMGENE variables.  
}
\usage{
mgVarPart(mg, X, categorical = NULL, categoryN = NULL,
    returnX = FALSE)
}
\arguments{
  \item{mg}{
}
  \item{X}{
}
  \item{categorical}{
}
  \item{categoryN}{
}
  \item{returnX}{
}
}
\details{
Under development
}

\author{
Pedro Peres-Neto (peres-neto.pedro@uqam.ca)\cr
Paul Galpern (pgalpern@ucalgary.ca)
}
